[torch/elastic] Introduce `DefaultRendezvousHandler` and `RendezvousBackend`.
This diff introduces the `DefaultRendezvousHandler` type as a stub implementation and its accompanying `RendezvousBackend` interface.

`DefaultRendezvousHandler` is intended to be a backend-agnostic type that will contain the core (bulk) logic of rendezvous handling. Any backend specific operation will be delegated to a concrete subclass of `RendezvousBackend` (e.g. `C10dRendezvousBackend` - see D27654492) that is passed as a constructor argument to `DefaultRendezvousHandler`.

Differential Revision: [D27654478](https://our.internmc.facebook.com/intern/diff/D27654478/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D27654478/)!

ghstack-source-id: 126092124
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55635