Fix complex mean and reduction tests not being run
Mean is broken for complex types, since #53218 it's now allocating the result
as a real tensor which discards the imaginary component. This wasn't picked up
in testing because `_test_dim_ops` tests are defined as closures inside of
`_test_dim_ops` instead of as methods on the test class. The result is, they
never get run.

For best results, view diff with "Hide whitespace changes".

ghstack-source-id: a81f040af88e29f16d6fc8a63c4239774a88a1d1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55640